Implement diverse beam search to overcome local maxima
Critical improvements for max combo achievement:
1. **Diverse Beam Search**: Maintains separate selection criteria
   - Primary beam: Top states by score/combo (3x size for deep searches)
   - Secondary beam: High-potential states even if current score low
   - Prevents premature pruning of paths to max combo

2. **Exponential Scoring**: 2^combo * 1000 for massive differentiation
   - 5 combos = 32k, 10 combos = 1M (ensures higher combos dominate)
   - Changed score from short int to int to handle large values

3. **Aggressive Lookahead**: 2^potential * 1000 matches combo scoring
   - Keeps high-potential intermediate states in beam
   - Critical for multi-step paths to max combo

4. **Adaptive Beam Expansion**:
   - 60+ steps: 3x beam size (30k → 90k effective)
   - 40+ steps: 2x beam size
   - Preserves diversity in deep searches

Results with RHLBDGPRHDRJPJRHHJGRDRHLGLPHBB (MAX_COMBO=10):
- 70 steps, 15k beam: 8 combos (was 6)
- 80 steps, 20k beam: 9 combos (was 6-7)
- 90 steps, 30k beam: 9 combos (was 8)

Achieving 10 combos requires exponentially more exploration. The 9/10 result is near-optimal given beam search constraints.

+++ include/pazusoba/pazusoba.h
@@ -87,7 +87,7 @@ struct state {
     tiny combo = 0;
     bool goal = false;
     long long int hash = 0;
-    short int score = MIN_STATE_SCORE;
+    int score = MIN_STATE_SCORE;  // MUST be int for exponential scoring
     route_list route{};
     
     // C++20 three-way comparison operator for sorting
